+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - 2025-01-08T11:15:00: Added MongoDB entity models
 - 2025-01-15T16:45:00: Fixed local mongo URI properties
 - 2025-01-22T10:20:00: Optimized imports in DeepSeekDBModel
+- 2025-02-05T09:00:00: Created ModelService